% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxbootR-package.R
\docType{package}
\name{maxbootR-package}
\alias{maxbootR}
\alias{maxbootR-package}
\title{maxbootR: Efficient Bootstrap Methods for Block Maxima}
\description{
Implements state-of-the-art block bootstrap methods for extreme value statistics based on block maxima. Includes disjoint blocks, sliding blocks, and a generalized k-max transformation with circular recombination. Fast C++ backends (via Rcpp) ensure scalability for large time series. The methods follow the approach described in the paper ...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://arxiv.org/abs/2409.05529}
  \item \url{https://github.com/torbenstaud/maxbootR,}
}

}
\author{
\strong{Maintainer}: Torben Staud \email{torben.staud@gmail.com}

}
\keyword{internal}
