% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxbootR-package.R
\docType{package}
\name{maxbootR-package}
\alias{maxbootR}
\alias{maxbootR-package}
\title{maxbootR: Efficient Bootstrap Methods for Block Maxima}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implements state-of-the-art block bootstrap methods for extreme value statistics based on block maxima. Includes disjoint blocks, sliding blocks, relying on a circular transformation of blocks. Fast C++ backends (via Rcpp) ensure scalability for large time series.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://torbenstaud.github.io/maxbootR/}
  \item \url{https://github.com/torbenstaud/maxbootR}
  \item Report bugs at \url{https://github.com/torbenstaud/maxbootR/issues}
}

}
\author{
\strong{Maintainer}: Torben Staud \email{torben.staud@gmail.com} [copyright holder]

}
\keyword{internal}
